requirements.txt
flask>=3.0.0
flask-login>=0.6.3
pillow>=10.1.0
rembg>=2.0.50
firebase-admin>=6.4.0
werkzeug>=3.0.1
